+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Adds: normalzied all basic raw HTML elements [#103](https://github.com/Pageworks/papertrain/issues/103)
 - Adds: 404 page uses custom single and Complex Content builder
 - Adds: chrome viewport height fix [#87](https://github.com/Pageworks/papertrain/issues/87)
+- Adds: application now tracks touch elements for custom `has-touch` status class [#84](https://github.com/Pageworks/papertrain/issues/84)
 
 ### Fixes